+++ packages/rulesets/src/asyncapi/index.ts
@@ -1,4 +1,4 @@
-import { aas2_0, aas2_1, aas2_2, aas2_3 } from '@stoplight/spectral-formats';
+import { aas2_0, aas2_1, aas2_2, aas2_3, aas2_4 } from '@stoplight/spectral-formats';
 import {
   truthy,
   pattern,
@@ -8,13 +8,20 @@ import {
   alphabetical,
 } from '@stoplight/spectral-functions';
 
+import asyncApi2ChannelParameters from './functions/asyncApi2ChannelParameters';
+import asyncApi2ChannelServers from './functions/asyncApi2ChannelServers';
 import asyncApi2DocumentSchema from './functions/asyncApi2DocumentSchema';
+import asyncApi2MessageExamplesValidation from './functions/asyncApi2MessageExamplesValidation';
+import asyncApi2OperationIdUniqueness from './functions/asyncApi2OperationIdUniqueness';
 import asyncApi2SchemaValidation from './functions/asyncApi2SchemaValidation';
 import asyncApi2PayloadValidation from './functions/asyncApi2PayloadValidation';
+import asyncApi2ServerVariables from './functions/asyncApi2ServerVariables';
+import { uniquenessTags } from '../shared/functions';
+import asyncApi2Security from './functions/asyncApi2Security';
 
 export default {
   documentationUrl: 'https://meta.stoplight.io/docs/spectral/docs/reference/asyncapi-rules.md',
-  formats: [aas2_0, aas2_1, aas2_2, aas2_3],
+  formats: [aas2_0, aas2_1, aas2_2, aas2_3, aas2_4],
   rules: {
     'asyncapi-channel-no-empty-parameter': {
       description: 'Channel path must not have empty parameter substitution pattern.',
@@ -53,6 +60,28 @@ export default {
         functionOptions: {
           notMatch: '.+\\/$',
         },
+      },
+    },
+    'asyncapi-channel-parameters': {
+      description: 'Channel parameters must be defined and there must be no redundant parameters.',
+      message: '{{error}}',
+      severity: 'error',
+      type: 'validation',
+      recommended: true,
+      given: ['$.channels.*', '$.components.channels.*'],
+      then: {
+        function: asyncApi2ChannelParameters,
+      },
+    },
+    'asyncapi-channel-servers': {
+      description: 'Channel servers must be defined in the "servers" object.',
+      message: '{{error}}',
+      severity: 'error',
+      type: 'validation',
+      recommended: true,
+      given: '$',
+      then: {
+        function: asyncApi2ChannelServers,
       },
     },
     'asyncapi-headers-schema-type-object': {
@@ -142,6 +171,31 @@ export default {
         function: truthy,
       },
     },
+    'asyncapi-message-examples': {
+      description: 'Examples of message object should follow by "payload" and "headers" schemas.',
+      message: '{{error}}',
+      severity: 'error',
+      type: 'validation',
+      recommended: true,
+      given: [
+        // messages
+        '$.channels.*.[publish,subscribe].message',
+        '$.channels.*.[publish,subscribe].message.oneOf.*',
+        '$.components.channels.*.[publish,subscribe].message',
+        '$.components.channels.*.[publish,subscribe].message.oneOf.*',
+        '$.components.messages.*',
+        // message traits
+        '$.channels.*.[publish,subscribe].message.traits.*',
+        '$.channels.*.[publish,subscribe].message.oneOf.*.traits.*',
+        '$.components.channels.*.[publish,subscribe].message.traits.*',
+        '$.components.channels.*.[publish,subscribe].message.oneOf.*.traits.*',
+        '$.components.messages.*.traits.*',
+        '$.components.messageTraits.*',
+      ],
+      then: {
+        function: asyncApi2MessageExamplesValidation,
+      },
+    },
     'asyncapi-operation-description': {
       description: 'Operation "description" must be present and non-empty string.',
       recommended: true,
@@ -150,6 +204,16 @@ export default {
       then: {
         field: 'description',
         function: truthy,
+      },
+    },
+    'asyncapi-operation-operationId-uniqueness': {
+      description: '"operationId" must be unique across all the operations.',
+      severity: 'error',
+      recommended: true,
+      type: 'validation',
+      given: '$',
+      then: {
+        function: asyncApi2OperationIdUniqueness,
       },
     },
     'asyncapi-operation-operationId': {
@@ -161,6 +225,20 @@ export default {
       then: {
         field: 'operationId',
         function: truthy,
+      },
+    },
+    'asyncapi-operation-security': {
+      description: 'Operation have to reference a defined security schemes.',
+      message: '{{error}}',
+      severity: 'error',
+      type: 'validation',
+      recommended: true,
+      given: '$.channels[*][publish,subscribe].security.*',
+      then: {
+        function: asyncApi2Security,
+        functionOptions: {
+          objectType: 'Operation',
+        },
       },
     },
     'asyncapi-parameter-description': {
@@ -282,6 +360,17 @@ export default {
         function: asyncApi2DocumentSchema,
       },
     },
+    'asyncapi-server-variables': {
+      description: 'Server variables must be defined and there must be no redundant variables.',
+      message: '{{error}}',
+      severity: 'error',
+      type: 'validation',
+      recommended: true,
+      given: ['$.servers.*', '$.components.servers.*'],
+      then: {
+        function: asyncApi2ServerVariables,
+      },
+    },
     'asyncapi-server-no-empty-variable': {
       description: 'Server URL must not have empty variable substitution pattern.',
       recommended: true,
@@ -315,6 +404,20 @@ export default {
         function: pattern,
         functionOptions: {
           notMatch: 'example\\.com',
+        },
+      },
+    },
+    'asyncapi-server-security': {
+      description: 'Server have to reference a defined security schemes.',
+      message: '{{error}}',
+      severity: 'error',
+      type: 'validation',
+      recommended: true,
+      given: '$.servers.*.security.*',
+      then: {
+        function: asyncApi2Security,
+        functionOptions: {
+          objectType: 'Server',
         },
       },
     },
@@ -356,6 +459,40 @@ export default {
         functionOptions: {
           keyedBy: 'name',
         },
+      },
+    },
+    'asyncapi-tags-uniqueness': {
+      description: 'Each tag must have a unique name.',
+      message: '{{error}}',
+      severity: 'error',
+      recommended: true,
+      type: 'validation',
+      given: [
+        // root
+        '$.tags',
+        // operations
+        '$.channels.*.[publish,subscribe].tags',
+        '$.components.channels.*.[publish,subscribe].tags',
+        // operation traits
+        '$.channels.*.[publish,subscribe].traits.*.tags',
+        '$.components.channels.*.[publish,subscribe].traits.*.tags',
+        '$.components.operationTraits.*.tags',
+        // messages
+        '$.channels.*.[publish,subscribe].message.tags',
+        '$.channels.*.[publish,subscribe].message.oneOf.*.tags',
+        '$.components.channels.*.[publish,subscribe].message.tags',
+        '$.components.channels.*.[publish,subscribe].message.oneOf.*.tags',
+        '$.components.messages.*.tags',
+        // message traits
+        '$.channels.*.[publish,subscribe].message.traits.*.tags',
+        '$.channels.*.[publish,subscribe].message.oneOf.*.traits.*.tags',
+        '$.components.channels.*.[publish,subscribe].message.traits.*.tags',
+        '$.components.channels.*.[publish,subscribe].message.oneOf.*.traits.*.tags',
+        '$.components.messages.*.traits.*.tags',
+        '$.components.messageTraits.*.tags',
+      ],
+      then: {
+        function: uniquenessTags,
       },
     },
     'asyncapi-tags': {