+++ packages/cli/CHANGELOG.md
@@ -1,3 +1,24 @@
+# [@stoplight/spectral-cli-v6.4.2](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-cli-v6.4.1...@stoplight/spectral-cli-v6.4.2) (2022-07-21)
+
+
+### Bug Fixes
+
+* **core:** fix 'resolved vs unresolved' json path mapping ([#2202](https://github.com/stoplightio/spectral/issues/2202)) ([157ec59](https://github.com/stoplightio/spectral/commit/157ec592d8b3276094284fead7a08541b3f46f61))
+
+# [@stoplight/spectral-cli-v6.4.1](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-cli-v6.4.0...@stoplight/spectral-cli-v6.4.1) (2022-06-01)
+
+
+### Bug Fixes
+
+* **cli:** do not show 'or higher' if severity equals error ([#2172](https://github.com/stoplightio/spectral/issues/2172)) ([f31ec63](https://github.com/stoplightio/spectral/commit/f31ec636c912f3c9a53672e87a13ad724921b902))
+
+# [@stoplight/spectral-cli-v6.4.0](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-cli-v6.3.0...@stoplight/spectral-cli-v6.4.0) (2022-05-12)
+
+
+### Features
+
+* **cli:** sort linting results alphabetically ([#2147](https://github.com/stoplightio/spectral/issues/2147)) ([84d48cf](https://github.com/stoplightio/spectral/commit/84d48cf5e02780f0cbb9ae9074c03a618c2bc462))
+
 # [@stoplight/spectral-cli-v6.3.0](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-cli-v6.2.1...@stoplight/spectral-cli-v6.3.0) (2022-03-03)